update README, use www.zoom.us and use *_mac by default
This patch allows the application to run by default by simply cloning and
following the instructions. It uses www.zoom.us instead of dev.zoom.us which is
not available to external users.

The patch also uses the *_mac files by defaults since they arelady work. And it
also updates the README.md since all the steps were already performed in the
postinstall script (one fix was also needed here).
